Adds a main- form
I have no idea what this actually does. Maybe this is needed to get the app to run on Heroku? I'm switching between the Heroku Shouter tutorial (uses an older version of Compojure but has PostgreSQL setup and cleaner code) and Mark Gandolfo's blog tutorial (newer version of Compojure, but the code isn't very DRY and he uses MySQL for persistence) to try to get the bits I want, but I don't know enough yet to know if I'm duplicating things unnecessarily. Just trying to get this to work for now.